Update the changes from #409 and merge them. Use this new FileObjectPosix to implement the socket._fileobject where possible. Fixes #409.
